Finalizando Associado e Equipamento

+++ src/main/java/com/anpede/services/AssociadoService.java
@@ -5,13 +5,17 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.anpede.dto.AssociadoDTO;
 import com.anpede.entities.Associado;
 import com.anpede.repositories.AssociadoRepository;
-import com.anpede.services.exceptions.EntityNotFoundException;
+import com.anpede.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 @Service
@@ -48,51 +52,72 @@ public class AssociadoService {
 	}
 
 	@Transactional(readOnly = true)
+	public Page<AssociadoDTO> findAllPaged(Pageable pageable) {		
+		Page<Associado> list = repository.findAll(pageable);		
+		return null;
+	}
+	
+	@Transactional(readOnly = true)
 	public AssociadoDTO findById(Long id) {
 		Optional<Associado> obj = repository.findById(id);
 		Associado a = obj.get();
 		return new AssociadoDTO(a);
 	}
 
+	@Transactional
 	public AssociadoDTO insert(AssociadoDTO dto) {
 		Associado entity = new Associado();
+		converterDtoEmEntidade(dto, entity);
+			
+		entity = repository.save(entity);
+		return new AssociadoDTO(entity);
+	}
+
+	private void converterDtoEmEntidade(AssociadoDTO dto, Associado entity) {
 		entity.setNome(dto.getNome());
 		entity.setCPF(dto.getCPF());
 		entity.setDataNascimento(dto.getDataNascimento());
 		entity.setTelefone(dto.getTelefone());
 		entity.setEmail(dto.getEmail());
 		entity.setEndereco(dto.getEndereco());
-			
-		entity = repository.save(entity);
-		return new AssociadoDTO(entity);
 	}
 
 	@Transactional
 	public AssociadoDTO update(Long id, AssociadoDTO dto) {
 		try {
 			
-			Associado entity = repository.getReferenceById(id);
-			
-			entity.setNome(dto.getNome());
-			entity.setCPF(dto.getCPF());
-			entity.setDataNascimento(dto.getDataNascimento());
-			entity.setTelefone(dto.getTelefone());
-			entity.setEmail(dto.getEmail());
-			entity.setEndereco(dto.getEndereco());
-				
+			Associado entity = repository.getReferenceById(id);			
+			converterDtoEmEntidade(dto, entity);				
 			entity = repository.save(entity);
 			return new AssociadoDTO(entity);
 			
-		} catch(jakarta.persistence.EntityNotFoundException e) {
-			throw new EntityNotFoundException("O recurso com o ID "+id
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("O recurso com o ID "+id
 					+" não foi localizado");
 		}
 	}
-	
-	//Implementar o tratamento de exceções para a deleção de registros
+		
 	public void delete(Long id) {
-		repository.deleteById(id);		
+		/*try {
+			repository.deleteById(id);
+		} catch(EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException("O recurso com o ID "+id
+					+" não foi localizado");
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Violação de Integride, "
+					+ "você não pode excluir este arquivo.");
+		}*/
+		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("O recurso com o ID "+id
+					+" não foi localizado");
+		}
+		repository.deleteById(id);
 	}
+
+
+
+
